moved main declarations to package json
