--- index.js
@@ -1,9 +1,0 @@
-const adapter= require('@qavajs/playwright-runner-adapter');
-const worlds = require('./lib/QavajsPlaywrightWdioWorld');
-const { locator } = require('./lib/pageObject');
-
-module.exports = {
-    ...adapter,
-    ...worlds,
-    locator
-}
