+++ src/index.ts
@@ -1,6 +1,7 @@
-export * from './src/pageObject';
 export * from '@qavajs/playwright-runner-adapter';
 export * from '@qavajs/playwright-wdio-fixtures';
-export * from './src/QavajsPlaywrightWdioWorld';
+export * from './QavajsPlaywrightWdioWorld';
 export * from 'webdriverio';
-export * from './src/types'
+export * from './types'
+export * from './pageObject';
+
