+++ test-e2e/memory/index.ts
@@ -1,4 +1,4 @@
-import { resolve } from 'path';
+import { resolve } from 'node:path';
 import { Response } from '@playwright/test';
 const file = (path: string) => `file://${path}`
 export default class Memory {
